+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - Option to upload pre-sell images [#31](https://github.com/BIEMAX/quote-manager/issues/31)
 - Implement function to delete defaulter [#48](https://github.com/BIEMAX/quote-manager/issues/48)
 - Implement tags menu [#49](https://github.com/BIEMAX/quote-manager/issues/49)
+- Implement new options to register new presell
 
 #### Fix
 - Fix `Filter Panel` component not showing properly the components added in template slot